Upgrade to Laravel 8, Move models into models/, Deleted controllers from old php site
