--- app/Http/Controllers/AppController.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-use App\UserCollection ;
-use App\YouTubeVideo;
-
-class AppController extends Controller
-{
-}
