+++ app/Models/GlobalQueue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Carbon;
 use App\Media;
