+++ app/Models/MediaTempItem.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use App\Media;
 use App\UserMedia;
