+++ app/Models/Theme.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 class Theme {
     public static $themes = [];
