+++ app/Models/User.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use DB;
 use Auth;
 use Hash;
-use App\Follow;
 use App\Theme;
+use App\Follow;
 use App\UserLike;
 use App\YouTubeVideo;
 use Laravel\Sanctum\HasApiTokens;
 use Illuminate\Notifications\Notifiable;
-// use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 
 class User extends Authenticatable
 {
-  use HasApiTokens, Notifiable;
+  use HasFactory, HasApiTokens, Notifiable;
 
   /**
    * The attributes that are mass assignable.
