+++ app/Models/YouTubeVideo.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Auth;
 use App\UserCollection;
