+++ composer.json
@@ -5,18 +5,18 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.0.0",
+        "php": ">=7.3.0",
         "asvae/laravel-api-tester": "^2.0",
         "fideloper/proxy": "~4.0",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.2",
         "jwilsson/spotify-web-api-php": "2.12.0",
         "ksubileau/color-thief-php": "^1.4",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/helpers": "^1.2",
         "laravel/sanctum": "^2.8",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
+        "laravel/ui": "^3.0",
         "madcoda/php-youtube-api": "^1.2",
         "rap2hpoutre/laravel-log-viewer": "^1.6",
         "spatie/laravel-honeypot": "^2.2"
@@ -26,7 +26,7 @@
         "filp/whoops": "~2.0",
         "fzaninotto/faker": "~1.4",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "files": [
