+++ public/index.php
@@ -9,6 +9,10 @@
 
 define('LARAVEL_START', microtime(true));
 
+if (file_exists(__DIR__ . '/../storage/framework/maintenance.php')) {
+    require __DIR__ . '/../storage/framework/maintenance.php';
+}
+
 /*
 |--------------------------------------------------------------------------
 | Register The Auto Loader
@@ -36,9 +40,6 @@ require __DIR__.'/../vendor/autoload.php';
 */
 
 $app = require_once __DIR__.'/../bootstrap/app.php';
-
-//Load Global Functions
-require_once __DIR__.'/../app/Global.php';
 
 /*
 |--------------------------------------------------------------------------
